requirements.txt
sh >= 1.12
python-box >= 3.1
